requirements.txt
asgiref==3.6.0
better-profanity==0.7.0
boto3==1.26.54
botocore==1.29.54
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.52.0
django-ckeditor==6.5.1
django-countries==7.5
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-maintenance-mode==0.18.0
django-phonenumbers==1.0.1
django-recaptcha==3.0.0
django-social-share==2.3.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
phonenumbers==8.13.4
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-fsutil==0.9.3
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
